Update Sidebar and HelpWindow
Introduce UI updates to Sidebar by integrating icons in order to
ease UX by allowing for visual navigation of WhoAreYouAgain?

Additionally, styled and formatted HelpWindow in a more readable
and consistent theme to that of MainWindow. This should help
users obtain easy to read, digestible documentation which not only
makes sense but looks good!

To do tomorrow:
- Fill in content according to UG